manager(miuix): drop parallel UiState/Utils support files, reuse M3 types and utils
