+++ manager/app/src/main/java/com/resukisu/resukisu/ui/screen/about/AboutMiuix.kt
@@ -474,3 +474,28 @@ private fun AboutContent(
         }
     }
 }
+
+data class AboutScreenActions(
+    val onBack: () -> Unit,
+    val onOpenLink: (String) -> Unit,
+    val onOpenLicense: () -> Unit = {},
+)
+
+data class LinkInfo(
+    val fullText: String,
+    val url: String
+)
+
+fun extractLinks(html: String): List<LinkInfo> {
+    val lines = html.split("<br/>", "<br>", "\n")
+    val result = mutableListOf<LinkInfo>()
+    for (line in lines) {
+        val spanned: android.text.Spanned = android.text.Html.fromHtml(line, android.text.Html.FROM_HTML_MODE_LEGACY)
+        val spans = spanned.getSpans(0, spanned.length, android.text.style.URLSpan::class.java)
+        val text = spanned.toString().trim()
+        for (span in spans) {
+            result.add(LinkInfo(text, span.url))
+        }
+    }
+    return result
+}
