--- manager/app/src/main/java/com/resukisu/resukisu/ui/screen/about/AboutUiState.kt
@@ -1,10 +1,0 @@
-package com.resukisu.resukisu.ui.screen.about
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-data class AboutScreenActions(
-    val onBack: () -> Unit,
-    val onOpenLink: (String) -> Unit,
-    val onOpenLicense: () -> Unit = {},
-)
