+++ manager/app/src/main/java/com/resukisu/resukisu/ui/screen/flash/FlashMiuix.kt
@@ -203,3 +203,11 @@ private fun TopBar(
         )
     }
 }
+
+data class FlashScreenActions(
+    val onBack: () -> Unit,
+    val onSaveLog: () -> Unit,
+    val onReboot: () -> Unit,
+    val onConfirmJailbreakWarning: () -> Unit,
+    val onDismissJailbreakWarning: () -> Unit,
+)
