--- manager/app/src/main/java/com/resukisu/resukisu/ui/screen/flash/FlashUiState.kt
@@ -1,13 +1,0 @@
-package com.resukisu.resukisu.ui.screen.flash
-
-import androidx.compose.runtime.Immutable
-import com.resukisu.resukisu.ui.screen.FlashingStatus
-
-@Immutable
-data class FlashScreenActions(
-    val onBack: () -> Unit,
-    val onSaveLog: () -> Unit,
-    val onReboot: () -> Unit,
-    val onConfirmJailbreakWarning: () -> Unit,
-    val onDismissJailbreakWarning: () -> Unit,
-)
