--- manager/app/src/main/java/com/resukisu/resukisu/ui/screen/flash/UninstallType.kt
@@ -1,30 +1,0 @@
-// UninstallType from tiann/KernelSU (github.com/tiann/KernelSU), used by the ported
-// Miuix uninstall dialog. GPL-3.0. See docs/ATTRIBUTION.md.
-package com.resukisu.resukisu.ui.screen.flash
-
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Adb
-import androidx.compose.material.icons.rounded.DeleteForever
-import androidx.compose.material.icons.rounded.RemoveModerator
-import androidx.compose.material.icons.rounded.RestartAlt
-import androidx.compose.ui.graphics.vector.ImageVector
-import com.resukisu.resukisu.R
-
-enum class UninstallType(val icon: ImageVector, val title: Int, val message: Int) {
-    TEMPORARY(
-        Icons.Rounded.RemoveModerator,
-        R.string.settings_uninstall_temporary,
-        R.string.settings_uninstall_temporary_message
-    ),
-    PERMANENT(
-        Icons.Rounded.DeleteForever,
-        R.string.settings_uninstall_permanent,
-        R.string.settings_uninstall_permanent_message
-    ),
-    RESTORE_STOCK_IMAGE(
-        Icons.Rounded.RestartAlt,
-        R.string.settings_restore_stock_image,
-        R.string.settings_restore_stock_image_message
-    ),
-    NONE(Icons.Rounded.Adb, 0, 0)
-}
