--- manager/app/src/main/java/com/resukisu/resukisu/ui/screen/home/HomeUtils.kt
@@ -1,30 +1,0 @@
-package com.resukisu.resukisu.ui.screen.home
-
-import android.content.Context
-import androidx.compose.runtime.Immutable
-import androidx.core.content.pm.PackageInfoCompat
-
-@Immutable
-data class ManagerVersion(
-    val versionName: String,
-    val versionCode: Long
-)
-
-@Immutable
-data class SystemInfo(
-    val kernelVersion: String,
-    val managerVersion: String,
-    val deviceModel: String,
-    val fingerprint: String,
-    val selinuxStatus: String,
-    val seccompStatus: Int
-)
-
-fun getManagerVersion(context: Context): ManagerVersion {
-    val packageInfo = context.packageManager.getPackageInfo(context.packageName, 0)!!
-    val versionCode = PackageInfoCompat.getLongVersionCode(packageInfo)
-    return ManagerVersion(
-        versionName = packageInfo.versionName!!,
-        versionCode = versionCode
-    )
-}
