--- manager/app/src/main/java/com/resukisu/resukisu/ui/screen/module/ModuleUiState.kt
@@ -1,61 +1,0 @@
-package com.resukisu.resukisu.ui.screen.module
-
-import android.net.Uri
-import androidx.compose.runtime.Immutable
-import com.resukisu.resukisu.ui.viewmodel.ModuleViewModel.ModuleInfo as Module
-import com.resukisu.resukisu.ui.viewmodel.ModuleViewModel.ModuleUpdateInfo
-import com.resukisu.resukisu.ui.component.SearchStatus
-
-sealed interface ModuleConfirmRequest {
-    data class Update(
-        val module: Module,
-        val downloadUrl: String,
-        val fileName: String,
-    ) : ModuleConfirmRequest
-
-    data class Uninstall(
-        val module: Module,
-    ) : ModuleConfirmRequest
-}
-
-@Immutable
-data class ModuleConfirmDialogState(
-    val request: ModuleConfirmRequest,
-    val title: String,
-    val content: String? = null,
-    val markdown: Boolean = false,
-    val html: Boolean = false,
-    val confirm: String? = null,
-    val dismiss: String? = null,
-)
-
-sealed interface ModuleEffect {
-    data class Toast(
-        val message: String,
-    ) : ModuleEffect
-
-    data class SnackBar(
-        val message: String,
-    ) : ModuleEffect
-}
-
-@Immutable
-data class ModuleActions(
-    val onRefresh: () -> Unit,
-    val onSearchStatusChange: (SearchStatus) -> Unit,
-    val onSearchTextChange: (String) -> Unit,
-    val onClearSearch: () -> Unit,
-    val onRequestUpdateConfirmation: (Module, ModuleUpdateInfo) -> Unit,
-    val onRequestUninstallConfirmation: (Module) -> Unit,
-    val onDismissConfirmRequest: () -> Unit,
-    val onConfirmUpdate: (ModuleConfirmRequest.Update) -> Unit,
-    val onOpenRepo: () -> Unit,
-    val onToggleSortActionFirst: () -> Unit,
-    val onToggleSortEnabledFirst: () -> Unit,
-    val onOpenWebUi: (Module) -> Unit,
-    val onToggleModule: (Module) -> Unit,
-    val onUninstallModule: (Module) -> Unit,
-    val onUndoUninstallModule: (Module) -> Unit,
-    val onOpenFlash: (List<Uri>) -> Unit,
-    val onExecuteModuleAction: (Module) -> Unit,
-)
