+++ manager/app/src/main/java/com/resukisu/resukisu/ui/screen/modulerepo/ModuleRepoMiuix.kt
@@ -1205,3 +1205,46 @@ fun ModuleRepoDetailScreenMiuix(
         }
     }
 }
+
+enum class RepoSort { UPDATED, CREATED, NAME, STARS }
+
+data class ModuleRepoDetailActions(
+    val onBack: () -> Unit,
+    val onOpenWebUrl: () -> Unit,
+    val onOpenUrl: (String) -> Unit,
+    val onInstallModule: (android.net.Uri) -> Unit,
+)
+
+@kotlinx.parcelize.Parcelize
+data class ReleaseAssetArg(
+    val name: String,
+    val downloadUrl: String,
+    val size: Long,
+    val downloadCount: Int
+) : android.os.Parcelable
+
+@kotlinx.parcelize.Parcelize
+data class ReleaseArg(
+    val tagName: String,
+    val name: String,
+    val publishedAt: String,
+    val assets: List<ReleaseAssetArg>,
+    val descriptionHTML: String
+) : android.os.Parcelable
+
+@kotlinx.parcelize.Parcelize
+data class AuthorArg(
+    val name: String,
+    val link: String,
+) : android.os.Parcelable
+
+@kotlinx.parcelize.Parcelize
+data class RepoModuleArg(
+    val moduleId: String,
+    val moduleName: String,
+    val authors: String,
+    val authorsList: List<AuthorArg>,
+    val latestRelease: String,
+    val latestReleaseTime: String,
+    val releases: List<ReleaseArg>
+) : android.os.Parcelable
